+++ qiskit_experiments/library/calibration/fine_amplitude.py
@@ -21,7 +21,7 @@ from qiskit.providers import Backend
 from qiskit.pulse.schedule import ScheduleBlock
 
 from qiskit_experiments.framework import BaseExperiment, Options
-from qiskit_experiments.library.calibration.analysis.fine_amplitude_analysis import (
+from qiskit_experiments.library.calibration.analysis.error_amplification_analysis import (
     ErrorAmplificationAnalysis,
 )
 from qiskit_experiments.exceptions import CalibrationError
